Add interrupt isolation feature for threads
Any thread can now be admitted with an interrupt priority class 0xP.
When the thread is running, it can only be interrupted by an interrupt
of vector 0xQd, where Q>P.    This is the TPR priority class (CR8).

Because Nautilus places timer, kick and APIC emergency interrupts in
class 0xf (vector=0xf0+), when a thread asks for class 0xe, it is
isolated from all interrupts except (normally) the scheduling interrupts
(timer and kick).

Signed-off-by: Kyle Hale <khale@cs.iit.edu>