Add interrupt thread model for interrupt isolation (optional)
When configured, a special periodic real time thread is created on
each core.  Except for scheduling-related interrupts, no other
interrupts are enabled except when this thread is running.   As
a result, interrupts are isolated to this "interrupt thread".
This means that all threads are insulated from interrupts.  This is
in contrast to the interrupt priority class model, where a thread
can elect to explicitly insulate itself. Enabling
the interrupt thread model disables the interrupt priority class
model.

It is also possible to allow interrupts while the idle thread is
running.

Signed-off-by: Kyle Hale <khale@cs.iit.edu>