Real mode interface capability (optional)
With this option enabled, it is possible to call to real mode
code from long mode.  This allows long mode code to make
use of features such as a VESA-compliant video cards.

Although the code supports a classic BIOS "int x" interface for real mode,
the framework should apply equally well for more complex real mode code.

Shell code added to make it easy to play with this interactively.

Signed-off-by: Kyle Hale <khale@cs.iit.edu>