Add optional GPIO device abstraction, implementation, and example code
This adds the abstraction of an optional general purpose input/output
device and a simple implementation using a parallel port card.

Using this feature, it is possible to instrument code for external
observation at a lower level than even serial output, plus it makes
it possible to do external timing studies via an oscilloscope
or logic anaylyzer.

Example code is given in the shell.

Shell has also been extended with an ioapic debugging command.

(Rebased commit to apply on top of LUA changes)

Signed-off-by: Kyle Hale <khale@cs.iit.edu>