buddy alloc: block order correctly adjusted when trimming
When a block was being allocated from a free region that is too big,
the allocator would trim off the extra space, but would not adjust
the order of the returned block. Functionally this was not an issue
since buddy_free uses the programmer specified order (rather than
the order embedded in the block structure). However this can produce
wacky results when you're trying to do accounting based on the
allocated blocks.