backspace support in shell/virtual consoles
This adds backspace support to VCs. Ideally this should be completely
independent of a VC, especially when the console is in raw mode.
We'll want to eventually change this so that there are callbacks to
the VC creator for keys that involve directly changing the state
of the internal line buffer (cooked) or moving around the cursor
etc. (raw). I've made a small change for the beginnings of this
by encapsulating callback functions passed to VC creation in
an ops structure.