Add thread reanimation (thread pools)
An exited (and joined) thread can now be reused.  The scheduler
provides an interface to search for such a thread based on the
required CPU and stack size.  The thread creation logic first
attempts to reanimate a thread via this interface to avoid
the memory allocations and other thread creation costs.

Note that now an exited thread may: (1) be reanimated as described
above, (2) reaped via an explicit reap request, (3) reaped via
a periodically run reaping thread, or (4) reaped on an initial
malloc() failure (which triggers a reap as a simple form of
garbage collection)

Signed-off-by: Kyle Hale <khale@cs.iit.edu>