Bug fix and sanity-check for kmem_free
This cleans up the semantics of the block_hash code, fixes a possible
race between a deallocation and allocation of a block hash entry,
and adds logic to kmem_free that can catch a race due to a double
free (in some cases).

Note that the block_hash allocation/deallocation code has the assumption
that the caller (e.g., ultimatly the caller of malloc/free) is correct.
The logic is designed to handle races on an entry between multiple allocations
and a single deallocation.  If for some reason we ever see multiple deallocations
for the same entry at once, all hell can break loose:
  a=malloc(); // uses entry 5, 5 points to metadata about a
  free(a);    // releases a and entry 5
  b=malloc(); // reuses entry 5, 5 now points to metadata about b
  free(a);    // releases *b* and entry 5

Signed-off-by: Kyle Hale <khale@cs.iit.edu>