Add tasks
This adds the ability to create tasks (callbacks, DPCs, ...).    A
task consists of a callback function and an argument.   A task is
produced, consumed, completed, and can optionally be waited on.

Task can be of known or unknown size.   Tasks of known size must never
block, while tasks of unknown size may block, but ideally should not.

Tasks of known size can be run by the scheduler when they will not
perturb the eager execution of any real-time thread.

Tasks of known or unknown size can be run by idle threads, higher
priority per-cpu task threads, in the context of a task wait, or
by other task pumps the user may introduce.

Execution of tasks in the idle threads has a subtle interaction
with real-time, and there are compile-time options to select
which interaction you want, if any.

Test routines are included.

Signed-off-by: Kyle Hale <khale@cs.iit.edu>