Add QEMU shutdown utility function
This adds support to explicitly request a shutdown from QEMU (but it should
work for BOCHS too). This is useful, e.g. if you want to run scripts with
Nautilus emulated and you don't want it hanging at the root shell. You can just
throw in a call to qemu_shutdown() at the end of init.c.

You may want to add the -device isa-debug-exit to your QEMU command line
invocation if you're running a relatively new version of QEMU.