Fix race condition on sleep queues (probably) when waker is in interrupt context
Keep interrupts off for the duration of going to sleep on a wait
queue.   Previously, we would keep interrupts off only to the
point at which we could disable preemption.  If, however, we want
interrupt context to wake a wait queue, this makes it possible
for our sleep call, while it is holding the wait queue lock, to
to be interrupted by an interrupt handler that will spin on
the wait queue lock, resulting in a deadlock.   This fix prevents
this race from occuring.

Caveats: The assumption here is that the caller understands
that going to sleep on a wait queue will result in interrupts
being enabled on return, if we sleep.   Going to sleep with
interrupts off (or a lock held!) is very bad form, so...

Signed-off-by: Kyle Hale <khale@cs.iit.edu>