Add message queues (and some cleanup)
This adds classic message queues.   These are named, searchable, and
can be dumped out from (for example) the shell.   Message queues are
refcounted.

This commit also updates the arch-specific init()s to initialize
message queues and semaphores.

It also tweaks the Kconfig menu structures to be a bit easier to
read/understand.  This includes a new heading, "Languages", which
currently only has LUA under it.   The idea is that these are built-in
languages.

Signed-off-by: Kyle Hale <khale@cs.iit.edu>