support for dynamic linking
This adds support for a runtime dynamic linker. The model here is that
the kernel symbol table will always be passed in as as multiboot2 module
(a binary blob). The linker will stash this symbol table for later use.
The user can then run programs by passing in a program as *another*
multiboot2 module (assumed to be a position-independent, shared object).
This program can only currently be invoked via the shell command "prog run."
There is only one program (for now). You can get information about the program
by running "prog info." Note that the program's symbolic references (to Nautilus
functions, other shared libs have not yet been attempted) will be resolved
eagerly when the program is invoked. This is equivalent to passing LD_BIND_NOW
to GNU ld.

Signed-off-by: Kyle Hale <khale@cs.iit.edu>