initial Clang/LLVM support
This adds the ability to compile Nautilus with Clang/LLVM if USE_CLANG=1
is passed to make. We still use GNU ld as the linker, although this will
hopefully change to incorporate LLVM's lld. This is experimental
support. If the user does nothing, this will default to using gcc. Here
are the following caveats when compiling with Clang

* must compile -O1 (O0 or O2 and above do not work)
* cannot have -Wno-frame-address
* cannot have -Wno-unused-but-set-variable
* cannot use -std=gnu99
* lots of warnings (e.g. empty loop bodies, packed fields etc. etc.)
* LLVM's linker doesn't work (probably insufficient linker script
support)
* Clang is a bit more aggressive with vectorization it seems
* we still use gcc for host programs
* this may be breaking interrupt handlers, as keyboard doesn't seem to
work

Note, this has only been tested on QEMU. It may break on real HW, so
use at your own risk