fix SSE alignment #GP
When we switch from the boot-time stack to the scheduler-allocated stack, we
drop the RSP pointer down by one word (so as not to push past the bottom of the
stack). This one word adjustment caused RSP (and thus later RBP) to be aligned
to an 8-byte (rather than 16-byte) boundary. SSE instructions involving memory
access (e.g. movaps) require that the memory operand be aligned on a 16-byte
boundary, so after the stack switch (both on BSP and APs) we'd see GPFs raised
when these instructions were executed.

This should remove the requirement that we compile with -mno-sse*.