do not panic on FPU exception + paranoid FPU init
This prevents Nautilus from panicing when it gets an FPU
exception (e.g. divide-by-zero, rounding, precisino etc). Rather,
it now prints a warning, clears the FPU exception flags, and continues.

This also is paranoid about how the FPU is initialized, banging
on the control register with the value that it *should* have been
loaded with after execution of FNINIT. A few utility functions for
manipulating FPU ctrl/status were added.