fix buddy system corruption on large memory machines
This fixes a particularly nasty bug in the buddy system's use of bitmap utility
functions. We encountered this issue after installing new DIMMs on our dev
server, bringing what was a 16GB machine up to one with 48GB. Suddenly the kernel
started to mysteriously hang early on.

The culprit turned out to be the buddy allocator's use of bitmap_zero().  The
buddy system maintains a bitmap with a bit for each (minimum sized) block it
manages. By default in Nautilus, the minimum block size is 32B. Nautilus keeps
a buddy allocator on top of each memory pool (roughly corresponding to a NUMA
zone).  The number of bits in the bitmap are therefore equal to the number of
minimum sized blocks in the pool (calculated as 1<<pool_order / (1<<
min_order)).

One of the first things the buddy system does is zero out this bitmap:

bitmap_zero(mp->tag_bits, mp->num_blocks);

Which, of course, *should* be fine.

*Until* you see that the second argument of bitmap_zero() is actually a plain
old int. This can represent up to +2^31-1, which translates into just below 2
billion blocks (should be plenty right?) before the int overflow. With the 32B
block size, this means we can have at most (2^31 - 1)(2^5) bytes (or just shy
of 64GB) before things start to break.  But our machine only has 48GB!  This
issue is that when we represent the size of the pool as a power of 2, we have
to round quite a bit. Unfortunately, we round *up* (this should probably
change, but I'm not adding that change here). 48GB will round up to exactly
64GB, which triggers the bug (because the bitmap_zero() function's underlying
memset will treat its length argument as an unsigned value), resulting in some
pretty serious memory corruption.  What this means is that any system with >
32GB of RAM would have triggered it as well.