add test matrix capability and associated scripts
With this addition, all tests to be automatically run are
added to the file .test-matrix.yml. To invoke all of those tests,
one uses scripts/run_tests.py, which scans that YAML file, and
for each test contained therein executes it with the indicated
kernel configurations. The success or failure of the test is determined
by the (non-zero) exit code produced by QEMU