Numerous fixes for compilation in Clang 6.0
These fixes allow the vast majority of NK to compile
using both GCC  (5.4.0 / Ubuntu 16 generic version) and
CLANG 6.0 (Ubuntu 16 version).  The only known exceptions are:

- OpenMP tests (for obvious reasons)
- LUA tests (script-generated code doesn't compile in either
  GCC or CLANG)
- Legion RT (not tested)

Note that the core kernel functionality works, with up to -O3
optimization in CLANG 3.6 and 6.0 (see previous commit).
This patch makes (most) of the rest of the world compile -
it doesn't test it.

Specific tweaks:

- realmode => alignment constraint to fix warning, hardcoded retf
    instruction since lvm-as doesnt do retf

- gc / boehm collector => warning / minor bugs if testing code not on
  added relevant include files since thread.h has changed slightly

- lua - warning fixes, minor bug in change to vc

- dtoa - warning fix

- vc - minor bug fix

- lwip - minor bug fix

- ipi - minor bug fix for when benchmark is enabled with filesystem

- gpio => minor bug if shell tests enabled

- mlx3 => warning fix

- vesa => minor bug

- ndpc => warning fix for test code

- nesl => minor bug and warning fixes