Add watchdog interface and mechanism
This is currently implemented using the i8254 PIT and the
first IOAPIC to send an NMI to cpu 0.   The watchdog is
petted by interrupt entry code.

Because the i8254 cannot manage longer intervals than about
54 ms, you need to be sure the watchdog is petted more frequently
than this.   Note that with the default 10 Hz timer rate, this is
not the case.