add idle thread for testing framework
If we're booting in "testing" mode, meaning we run all available
tests at boot, there's a chance we won't have a thread to switch
to if a tester creates a thread that blocks. This fixes that by
making sure there is an idle thread on core 0 to switch to. This
idle thread is *only* going to be present when the testing framework
is compiled in and tests are being run at boot-time. Performance
testing should be done via shell scripts etc.